Move Helsinki specific stuff to parkkihubi_hel
