+++ setup.cfg
@@ -23,7 +23,7 @@ filterwarnings =
 
 [isort]
 default_section=THIRDPARTY
-known_first_party=parkkihubi,parkings
+known_first_party=parkkihubi,parkings,parkkihubi_hel
 line_length=79
 multi_line_output=4
 not_skip=__init__.py
